refactor(ui): split PageHeader into per-variant components
