+++ apps/web/src/components/layout/page-header/DraftsHeader.tsx
@@ -1,0 +1,23 @@
+import { Link } from 'react-router-dom';
+import { Button } from '../../ui';
+import { IconPencil } from './icons';
+
+export function DraftsHeader() {
+  return (
+    <>
+      <div className="flex items-center gap-2 text-sm font-medium text-(--txt-secondary)">
+        <span className="flex size-5 items-center justify-center text-(--txt-icon-tertiary)">
+          <IconPencil />
+        </span>
+        Drafts
+      </div>
+      <div className="flex items-center gap-2">
+        <Link to="?create=1" className="no-underline">
+          <Button size="sm" className="gap-1.5 text-[13px] font-medium">
+            Draft a work item
+          </Button>
+        </Link>
+      </div>
+    </>
+  );
+}
