+++ apps/web/src/components/layout/page-header/ProjectDetailHeader.tsx
@@ -1,0 +1,38 @@
+import { ProjectIconDisplay } from '../../ProjectIconModal';
+import type { ProjectApiResponse } from '../../../api/types';
+import { IconSearch } from './icons';
+
+export function ProjectDetailHeader({
+  project,
+  title,
+}: {
+  workspaceSlug: string;
+  projectId: string;
+  project: ProjectApiResponse;
+  title: string;
+}) {
+  return (
+    <>
+      <div className="flex items-center gap-2 text-sm font-semibold text-(--txt-primary)">
+        <span className="flex size-5 shrink-0 items-center justify-center">
+          <ProjectIconDisplay
+            emoji={project.emoji}
+            icon_prop={project.icon_prop}
+            size={16}
+            className="leading-none"
+          />
+        </span>
+        {title}
+      </div>
+      <div className="flex items-center gap-2">
+        <button
+          type="button"
+          className="flex size-8 items-center justify-center rounded-md text-(--txt-icon-tertiary) hover:bg-(--bg-layer-transparent-hover) hover:text-(--txt-icon-secondary)"
+          aria-label="Search"
+        >
+          <IconSearch />
+        </button>
+      </div>
+    </>
+  );
+}
